+++ CHANGES.rst
@@ -5,11 +5,11 @@ Changelog
 2.4.0 (unreleased)
 ------------------
 
-- remove unused contentrules
-  [mamico]
-
 - Split booking notification gateways to 'App IO', 'Email' and 'SMS'
   [folix-01]
+
+- remove unused contentrules
+  [mamico]
 
 - Add booking reminder emails.
   [folix-01]